ex04 done

+++ django_3.0_sessions/sessions/ex/views.py
@@ -37,11 +37,13 @@ def handle_downvote(request, tip_id):
 
 def handle_delete(request, tip_id):
 	tip = get_object_or_404(Tip, id=tip_id)
-	tip.delete()
+	if tip.author == request.user or request.user.has_perm("delete_tip"):
+		tip.delete()
 
 def homepage(request):
 	if request.user.is_authenticated:
 		username = request.user.username
+		has_delete_perm = request.user.has_perm("ex.delete_tip")
 
 		if request.method == "POST":
 			if "create_tip" in request.POST:
@@ -51,7 +53,7 @@ def homepage(request):
 			elif "downvote" in request.POST:
 				handle_downvote(request, request.POST.get("tip_id"))
 			elif "delete" in request.POST:
-				handle_delete(request, request.POST.get("tip_id"))
+				to_del = handle_delete(request, request.POST.get("tip_id"))
 			return redirect("homepage")
 		else:
 			form = PostTipForm()
@@ -65,6 +67,7 @@ def homepage(request):
 			username = request.session["username"]
 
 		form = None
+		has_delete_perm = False
 	
 	tips = Tip.objects.all().order_by("-date")
 
@@ -74,7 +77,8 @@ def homepage(request):
 	context = {
 		"username": username,
 		"form": form,
-		"tips": tips
+		"tips": tips,
+		"has_delete_perm" : has_delete_perm
 	}
 	return render(request, "ex/homepage.html", context)
